# Performance improvements for `jz outdated` command